# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API untuk personal portfolio (projects, experiences, contact, dll).",
        "title": "Portfolio Backend API",
        "contact": {
            "name": "Fauzan",
            "email": "fauzan@pandi.id"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/experiences": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin-experiences"
                ],
                "summary": "List all experiences",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ExperienceResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin-experiences"
                ],
                "summary": "Create new experience",
                "parameters": [
                    {
                        "description": "Experience payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ExperienceCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExperienceResponse"
                        }
                    }
                }
            }
        },
        "/admin/experiences/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin-experiences"
                ],
                "summary": "Get experience detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExperienceResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin-experiences"
                ],
                "summary": "Update experience",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Experience ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ExperienceUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExperienceResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin-experiences"
                ],
                "summary": "Delete experience",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Experience ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ExperienceCreateRequest": {
            "type": "object",
            "required": [
                "company",
                "startDate",
                "title"
            ],
            "properties": {
                "company": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "description": "nullable",
                    "type": "string"
                },
                "highlights": {
                    "description": "teks bullet",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "isCurrent": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "startDate": {
                    "description": "\"2006-01-02\"",
                    "type": "string"
                },
                "tagIds": {
                    "description": "UUID string",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "handlers.ExperienceHighlightResponse": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
        "handlers.ExperienceResponse": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "description": "nullable",
                    "type": "string"
                },
                "highlights": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ExperienceHighlightResponse"
                    }
                },
                "id": {
                    "type": "string"
                },
                "isCurrent": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "startDate": {
                    "description": "\"2006-01-02\"",
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.TagResponse"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "handlers.ExperienceUpdateRequest": {
            "type": "object",
            "required": [
                "company",
                "startDate",
                "title"
            ],
            "properties": {
                "company": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "description": "nullable",
                    "type": "string"
                },
                "highlights": {
                    "description": "teks bullet",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "isCurrent": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "startDate": {
                    "description": "\"2006-01-02\"",
                    "type": "string"
                },
                "tagIds": {
                    "description": "UUID string",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "handlers.TagResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}